Indomethacin for Post-ERCP Pancreatitis

This dataset, post_ercp_pancreatitis_tbl_df, is a tibble containing results from a randomized, placebo-controlled, prospective 2-arm trial of rectal indomethacin (100 mg) versus placebo to prevent post-ERCP pancreatitis in 602 participants, as reported by Elmunzer, Higgins, et al. (2012) in the New England Journal of Medicine.

Usage

data(post_ercp_pancreatitis_tbl_df)

Format

A tibble with 602 observations and 33 variables:

id

Numeric subject identifier

site

Factor indicating study site (4 levels)

age

Numeric age of the participant

risk

Numeric risk score

gender

Factor indicating gender (2 levels)

outcome

Factor indicating study outcome (2 levels)

sod

Factor indicating presence of sphincter of Oddi dysfunction (2 levels)

pep

Factor indicating presence of post-ERCP pancreatitis (2 levels)

recpanc

Factor indicating recurrent pancreatitis (2 levels)

psphinc

Factor indicating pancreatic sphincterotomy (2 levels)

precut

Factor indicating precut sphincterotomy (2 levels)

difcan

Factor indicating difficult cannulation (2 levels)

pneudil

Factor indicating pneumatic dilation (2 levels)

amp

Factor indicating ampullary interventions (2 levels)

paninj

Factor indicating pancreatic injury (2 levels)

acinar

Factor indicating acinarization (2 levels)

brush

Factor indicating brushing procedures (2 levels)

asa81

Factor indicating ASA 81 mg use (3 levels)

asa325

Factor indicating ASA 325 mg use (3 levels)

asa

Factor indicating ASA status (3 levels)

prophystent

Factor indicating prophylactic stent placement (2 levels)

therastent

Factor indicating therapeutic stent use (2 levels)

pdstent

Factor indicating pancreatic duct stent (2 levels)

sodsom

Factor indicating somatostatin use for SOD (2 levels)

bsphinc

Factor indicating biliary sphincterotomy (2 levels)

bstent

Factor indicating biliary stent (2 levels)

chole

Factor indicating cholecystectomy (2 levels)

pbmal

Factor indicating presence of pancreaticobiliary malignancy (2 levels)

train

Factor indicating if performed by trainee (2 levels)

status

Factor indicating trial status (2 levels)

type

Factor indicating procedure type (4 levels)

rx

Factor indicating treatment group: placebo or indomethacin (2 levels)

bleed

Numeric bleeding indicator

Source

Data taken from the medicaldata package version 0.2.0.

Details

The dataset name has been kept as 'post_ercp_pancreatitis_tbl_df' to avoid confusion with other datasets in the R ecosystem. This naming convention helps distinguish this dataset as part of the DigestiveDataSets package and assists users in identifying its specific characteristics. The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified in any way.